refactor(#44) ディレクトリ名を修正
